# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a FWW Regulation Documentation API",
        "title": "FWW Regulation API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/check/blacklist": {
            "get": {
                "description": "Check KTP blacklist status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blacklist"
                ],
                "summary": "Check KTP blacklist status",
                "parameters": [
                    {
                        "description": "BlacklistRequest",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BlacklistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/check/dukcapil": {
            "get": {
                "description": "Check KTP status at Dukcapil",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dukcapil"
                ],
                "summary": "Check KTP status at Dukcapil",
                "parameters": [
                    {
                        "description": "DukcapilRequest",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DukcapilRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/check/pedulilindungi": {
            "get": {
                "description": "Check KTP status at Peduli Lindungi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "peduli lindungi"
                ],
                "summary": "Check KTP status at Peduli Lindungi",
                "parameters": [
                    {
                        "description": "PeduliLindungiRequest",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PeduliLindungiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.BlacklistRequest": {
            "type": "object",
            "properties": {
                "ktp": {
                    "type": "string"
                }
            }
        },
        "model.DukcapilRequest": {
            "type": "object",
            "properties": {
                "ktp": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "model.PeduliLindungiRequest": {
            "type": "object",
            "properties": {
                "ktp": {
                    "type": "string"
                }
            }
        },
        "model.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}